+++ src/Interpreter.cpp
@@ -5,7 +5,7 @@
 #include <memory>
 #include <sstream>
 #include <stdexcept>
-#include "Api.h"
+#include "Api.hpp"
 #include "utils.h"
 
 using namespace std;
